fix: move model to correct app
